Refactoring : LadderGame 에서 Line 클래스 추출
- 사다리를 만드는 행위를 Line 클래스를 만들어 책임과 역할을 주었습니다.
- LadderGame에서는 Player와 사다리 높이만 입력 받고 사다리를 만드는
  기능만 하게 했습니다.
